+++ guillotina/contrib/workflows/utility.py
@@ -8,6 +8,7 @@ from guillotina.contrib.workflows.interfaces import IWorkflow
 from guillotina.contrib.workflows.interfaces import IWorkflowBehavior
 from guillotina.contrib.workflows.interfaces import IWorkflowUtility
 from guillotina.event import notify
+from guillotina.events import ObjectModifiedEvent
 from guillotina.response import HTTPPreconditionFailed
 from guillotina.response import HTTPUnauthorized
 from guillotina.security.utils import apply_sharing
@@ -95,6 +96,7 @@ def create_workflow_factory(proto_name, proto_definition):
             workflow_behavior.register()
 
             await notify(WorkflowChangedEvent(self.context, self, action, comments))
+            await notify(ObjectModifiedEvent(self.context, payload={"review_state": new_state}))
             return history
 
     return Workflow